Refactor: user id 검사 추가

+++ backend/src/routers/v1/posts.ts
@@ -70,6 +70,21 @@ export default class PostsRouter {
     return response.json({ code: Enums.responseCode.SUCCESS, _id: comment._id });
   }
 
+  @Post('/:postID/likes')
+  @UseBefore(passport.authenticate('jwt-registered', { session: false }))
+  async postPostLike(@Req() request: Request, @Res() response: Response) {
+    const { postID } = request.params;
+    const { userID } = request.body;
+    if (!userID) {
+      throw new Error(Enums.error.WRONG_BODY_TYPE);
+    }
+    if (userID !== request.user?.userID) {
+      throw new Error(Enums.error.PERMISSION_DENIED);
+    }
+    const _id = await PostLikeService.createPostLike(userID, postID);
+    return response.json({ code: Enums.responseCode.SUCCESS, _id });
+  }
+
   @Post('/:postID/comments/:commentID/likes')
   @UseBefore(passport.authenticate('jwt-registered', { session: false }))
   async postCommentLike(@Req() request: Request, @Res() response: Response) {
@@ -86,36 +101,28 @@ export default class PostsRouter {
     return response.json({ code: Enums.responseCode.SUCCESS, _id });
   }
 
-  @Post('/:postID/likes')
+  @Delete('/:postID/likes/:likeID')
   @UseBefore(passport.authenticate('jwt-registered', { session: false }))
-  async postPostLike(@Req() request: Request, @Res() response: Response) {
-    const { postID } = request.params;
+  async deletePostLike(@Req() request: Request, @Res() response: Response) {
+    const { postID, likeID } = request.params;
     const { userID } = request.body;
-    if (!userID) {
-      throw new Error(Enums.error.WRONG_BODY_TYPE);
-    }
     if (userID !== request.user?.userID) {
       throw new Error(Enums.error.PERMISSION_DENIED);
     }
-    const _id = await PostLikeService.createPostLike(userID, postID);
-    return response.json({ code: Enums.responseCode.SUCCESS, _id });
+    await PostLikeService.removePostLike(userID, postID, likeID);
+    return response.json({ code: Enums.responseCode.SUCCESS });
   }
 
   @Delete('/:postID/comments/:commentID')
   @UseBefore(passport.authenticate('jwt-registered', { session: false }))
   async deleteComment(@Req() request: Request, @Res() response: Response) {
     const { postID, commentID } = request.params;
+    const { userID } = request.body;
+    if (userID !== request.user?.userID) {
+      throw new Error(Enums.error.PERMISSION_DENIED);
+    }
     await CommentService.existsComment(request.user!.userID, postID, commentID);
-    const result = await CommentService.removeComment(postID, commentID);
-    return response.json(result);
-  }
-
-  @Delete('/:postID/likes/:likeID')
-  @UseBefore(passport.authenticate('jwt-registered', { session: false }))
-  async deletePostLike(@Req() request: Request, @Res() response: Response) {
-    const { postID, likeID } = request.params;
-    const { userID } = request.user!;
-    await PostLikeService.removePostLike(userID, postID, likeID);
+    await CommentService.removeComment(postID, commentID);
     return response.json({ code: Enums.responseCode.SUCCESS });
   }
 
@@ -123,7 +130,10 @@ export default class PostsRouter {
   @UseBefore(passport.authenticate('jwt-registered', { session: false }))
   async deleteCommentLike(@Req() request: Request, @Res() response: Response) {
     const { postID, commentID, likeID } = request.params;
-    const { userID } = request.user!;
+    const { userID } = request.body;
+    if (userID !== request.user?.userID) {
+      throw new Error(Enums.error.PERMISSION_DENIED);
+    }
     await CommentService.existsComment(userID, postID, commentID);
     await CommentLikeService.removeCommentLike(commentID, likeID);
     return response.json({ code: Enums.responseCode.SUCCESS });
